Move DynamoDBItem to its own class
